# back Calculus/api/openapi.yml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
  description: API documentation for the application.

servers:
  - url: http://174.129.53.237:3000
    description: Local server

paths:
  /resultado:
    post:
      summary: Resultado
      operationId: addResultado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResultadoRequest"
      responses:
        "201":
          description: Resultado creado correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultadoResponse"
        "400":
          description: Error en la solicitud
        "500":
          description: Error en el servidor
      x-eov-operation-handler: resultados

  /accion:
    post:
      summary: Acciones
      operationId: addAccion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccionRequest"
      responses:
        "201":
          description: Accion creado correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccionResponse"
        "400":
          description: Error en la solicitud
        "500":
          description: Error en el servidor
      x-eov-operation-handler: accion

  /conversaciones:
    post:
      summary: Conversaciones
      operationId: Conversaciones
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversacionesRequest"
      responses:
        "200":
          description: Conversaciones obtenidas correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversacionesResponse"
        "500":
          description: Error en el servidor
      x-eov-operation-handler: conversaciones
  /analisis:
    post:
      summary: Analisis
      operationId: Analisis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalisisRequest"
      responses:
        "200":
          description: Analisis obtenido correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalisisResponse"
        "500":
          description: Error en el servidor
      x-eov-operation-handler: conversaciones

  /analisis2:
    post:
      summary: Analisis
      operationId: Analisis2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalisisRequest"
      responses:
        "200":
          description: Analisis obtenido correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalisisResponse"
        "500":
          description: Error en el servidor
      x-eov-operation-handler: conversaciones

  /admin:
    post:
      summary: Admin
      operationId: loginAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRequest"
      responses:
        "200":
          description: Login exitoso admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminResponse"
        "401":
          description: Credenciales inválidas
        "500":
          description: Error en el servidor
      x-eov-operation-handler: admin

  /login:
    post:
      summary: Login
      operationId: loginUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Credenciales inválidas
        "500":
          description: Error en el servidor
      x-eov-operation-handler: usuario

  /usuario:
    get:
      summary: Obtener usuarios
      operationId: obtenerUsuarios
      responses:
        "200":
          description: Usuarios obtenidos correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UsuariosResponse"
        "500":
          description: Error en el servidor
      x-eov-operation-handler: usuario
    post:
      summary: Usuario
      operationId: addUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioRequest"
      responses:
        "201":
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioResponse"
        "400":
          description: Error en la solicitud
        "500":
          description: Error en el servidor
      x-eov-operation-handler: usuario

  /usuario/{id}:
    delete:
      summary: Eliminar usuario
      operationId: eliminarUsuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Usuario eliminado correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioResponse"
        "404":
          description: Usuario no encontrado
        "500":
          description: Error en el servidor
      x-eov-operation-handler: usuario
    put:
      summary: Modificar usuario
      operationId: modificarUsuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioRequest"
      responses:
        "200":
          description: Usuario modificado correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioResponse"
        "400":
          description: Error en la solicitud
        "404":
          description: Usuario no encontrado
        "500":
          description: Error en el servidor
      x-eov-operation-handler: usuario

  /chat:
    post:
      summary: Chatbot
      operationId: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/chatRequest"
      responses:
        "200":
          description: Respuesta correcta del chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/chatResponse"
        "400":
          description: Error en la solicitud
        "500":
          description: Error en el servidor
      x-eov-operation-handler: chat

  /resultados-usuario:
    post:
      summary: Obtener resultados por usuario
      operationId: getResultadosUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                usuario_id:
                  type: integer
                  example: 1
      responses:
        "200":
          description: Resultados del usuario obtenidos correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResultadoResponse"
        "500":
          description: Error en el servidor
      x-eov-operation-handler: resultados

components:
  schemas:
    ResultadoRequest:
      type: object
      properties:
        usuario_id:
          type: integer
          example: 1
        resultado_alumno:
          type: string
          example: "Resultado del ejercicio"
        current_exercise:
          type: integer
          example: 1

    ResultadoResponse:
      type: object
      properties:
        usuario_id:
          type: integer
          example: 1
        resultado_alumno:
          type: string
          example: "Resultado del ejercicio"
        resultado_correcto:
          type: string
          example: "Resultado correcto del ejercicio"
        estado_ejercicio:
          type: string
          example: "malo"
        current_exercise:
          type: integer
          example: 1
        creado_en:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"

    AccionRequest:
      type: object
      properties:
        usuario_id:
          type: integer
          example: 1
        accion:
          type: string
          example: "click en el grafico"
        current_exercise:
          type: integer
          example: 1

    AccionResponse:
      type: object
      properties:
        accion:
          type: string
          example: "click en el grafico"
        usuario_id:
          type: integer
          example: 1
        current_exercise:
          type: integer
          example: 1
        timestamp:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"

    UsuariosResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Juan Perez"
        correo:
          type: string
          example: "juan@gmail.com"
        nickname:
          type: string
          example: "jperez"
        contrasena:
          type: string
          example: "password123"
        creado_en:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"

    AnalisisRequest:
      type: object
      properties:
        usuario_id:
          type: integer
          example: 1
    AnalisisResponse:
      type: object
      properties:
        analisis:
          type: string
          example: "Análisis de la conversación"
    ConversacionesRequest:
      type: object
      properties:
        usuario_id:
          type: integer
          example: 1
    ConversacionesResponse:
      type: array
      items:
        type: object
        properties:
          id:
            type: integer
            example: 1
          mensaje_usuario:
            type: string
            example: "Hola, ¿cómo estás?"
          respuesta_chatbot:
            type: string
            example: "Todo bien, ¿y tú?"
          fecha_hora:
            type: string
            format: date-time
            example: "2023-10-01T12:00:00Z"
          currentExercise:
            type: integer
            example: 1

    AdminRequest:
      type: object
      properties:
        nickname:
          type: string
          example: "admin"
        contrasena:
          type: string
          example: "admin123"

    AdminResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Admin"
        nickname:
          type: string
          example: "admin"
        contrasena:
          type: string
          example: "admin123"
        creado_en:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"

    LoginRequest:
      type: object
      properties:
        nickname:
          type: string
          example: "jperez"
        contrasena:
          type: string
          example: "password123"

    LoginResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Juan Perez"
        nickname:
          type: string
          example: "jperez"
       

    chatRequest:
      type: object
      properties:
        mensaje_usuario:
          type: string
          example: "Hola chabot , como estas?"
        conversationId:
          type: string
          nullable: true
          example: "conv_1743877684039"
        usuario_id:
          type: integer
          example: 1
        currentExercise:
          type: integer
          example: 1

    chatResponse:
      type: object
      properties:
        conversationId:
          type: string
          example: "conv_1743877684039"
        response:
          type: string
          example: "Todo bien y tu?"
        usuario_id:
          type: integer
          example: 1
        timestamp:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"
        currentExercise:
          type: integer
          example: 1
        collaborationPassed:
          type: boolean
          example: false
        PHASE_REACHED:
          type: boolean
          example: false

    UsuarioRequest:
      type: object
      properties:
        nombre:
          type: string
          example: "Juan Perez"
        correo:
          type: string
          example: "juan@gmail.com"
        nickname:
          type: string
          example: "jperez"
        contrasena:
          type: string
          example: "password123"

    UsuarioResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Juan Perez"
        correo:
          type: string
          example: "juan@gmail.com"
        nickname:
          type: string
          example: "jperez"
        creado_en:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"